Apply PR feedback
