+++ package.json
@@ -1747,6 +1747,7 @@
     "@kbn/es-archiver": "link:src/platform/packages/shared/kbn-es-archiver",
     "@kbn/es-snapshot-loader": "link:x-pack/platform/packages/shared/kbn-es-snapshot-loader",
     "@kbn/eslint-config": "link:packages/kbn-eslint-config",
+    "@kbn/eslint-config-security-tests": "link:packages/kbn-eslint-config-security-tests",
     "@kbn/eslint-plugin-alerting-v2": "link:packages/kbn-eslint-plugin-alerting-v2",
     "@kbn/eslint-plugin-disable": "link:packages/kbn-eslint-plugin-disable",
     "@kbn/eslint-plugin-eslint": "link:packages/kbn-eslint-plugin-eslint",
